[UPD] split logic in helper functions
