+++ src/utils/submitFindingsUtils/state.js
@@ -1,0 +1,21 @@
+const mdTemplate =
+"## Impact\nDetailed description of the impact of this finding.\n\n## Proof of Concept\nProvide direct links to all referenced code in GitHub. Add screenshots, logs, or any other relevant proof that illustrates the concept.\n\n## Tools Used\n\n## Recommended Mitigation Steps";
+
+export const initialState = {
+  contest: "",
+  sponsor: "",
+  repo: "",
+  title: "",
+  email: "",
+  handle: "",
+  polygonAddress: "",
+  risk: "",
+  details: mdTemplate,
+  qaGasDetails: "",
+  linesOfCode: [
+    {
+      id: Date.now(),
+      value: "",
+    },
+  ],
+};
